refactor(admin-artwork): separate the code of artwork gallery from the artist form
